Fix screen to buildings for case where screens are close to building or self intersects

+++ wps_scripts/src/main/groovy/org/noise_planet/noisemodelling/wps/Others_Tools/Screen_to_building.groovy
@@ -33,6 +33,19 @@ static Connection openGeoserverDataStoreConnection(String dbName) {
 }
 
 def run(input) {
+    // Get name of the database
+    String dbName = ""
+    if (input['databaseName']) {
+        dbName = input['databaseName'] as String
+    }
+    // Open connection
+    openGeoserverDataStoreConnection(dbName).withCloseable {
+        Connection connection ->
+            return [result: exec(connection, input)]
+    }
+}
+
+def exec(Connection connection, input) {
 
     String screen_table_name =  input['screenTableName']
     screen_table_name = screen_table_name.toUpperCase()
@@ -44,42 +57,56 @@ def run(input) {
 
     building_table_name = building_table_name.toUpperCase()
 
-    // Get name of the database
-    String dbName = ""
-    if (input['databaseName']) {
-        dbName = input['databaseName'] as String
+    Sql sql = new Sql(connection)
+
+    // Check for intersections between walls
+    int intersectingWalls = sql.firstRow("select count(*) interswalls from "+screen_table_name+" E1, "+screen_table_name+" E2 where E1.pk < E2.pk AND ST_Distance(E1.the_geom, E2.the_geom) < 0.2;")[0]
+    if(intersectingWalls > 0) {
+        return [tableNameCreated: String.format(Locale.ROOT, "%d screens are too close to each other (< 0.2m), please modify or delete them before including them in the simulation.", intersectingWalls)]
     }
 
-    // Open connection
-    openGeoserverDataStoreConnection(dbName).withCloseable { Connection connection ->
-        //Statement sql = connection.createStatement()
-        Sql sql = new Sql(connection)
+    if (input['buildingTableName']) {
 
-        sql.execute(String.format("DROP TABLE IF EXISTS SCREENS"))
-        sql.execute("create table SCREENS as select * from " + screen_table_name )
+        // Remove parts of the screen too close from buildings
+        double distance_truncate_screens = 2.0
+        // Find screen intersecting buildings
+        sql.execute("CREATE SPATIAL INDEX SCREEN_INDEX ON "+screen_table_name+"(the_geom)")
+        sql.execute("drop table if exists tmp_relation_screen_building;")
+        sql.execute("create table tmp_relation_screen_building as select b.pk as PK_building, s.pk as pk_screen" +
+                " from "+building_table_name+" b, "+screen_table_name+" s where b.the_geom && s.the_geom and" +
+                " ST_Distance(b.the_geom, s.the_geom) <= "+distance_truncate_screens)
+        // For intersecting screens, remove parts closer than distance_truncate_screens
+        sql.execute("drop table if exists tmp_screen_truncated;")
+        sql.execute("create table tmp_screen_truncated as select pk_screen, ST_DIFFERENCE(s.the_geom, " +
+                "ST_BUFFER(ST_ACCUM(b.the_geom), 2)) the_geom,s.height from tmp_relation_screen_building r, " +
+                building_table_name+" b, "+screen_table_name+" s WHERE PK_building = b.pk AND pk_screen = s.pk " +
+                "GROUP BY pk_screen, s.height;")
 
-        if (input['buildingTableName']) {
-            sql.execute("DROP TABLE IF EXISTS BUFFERED_SCREENS")
-            sql.execute("CREATE TABLE BUFFERED_SCREENS as select ST_BUFFER(sc.the_geom,0.1, 'join=bevel') the_geom,  HEIGHT HEIGHT from SCREENS sc")
-
-            sql.execute("DROP TABLE IF EXISTS BUILDINGS_SCREENS")
-            sql.execute("CREATE TABLE BUILDINGS_SCREENS as select ST_SimplifyPreserveTopology(the_geom,0.1) the_geom, HEIGHT from BUFFERED_SCREENS sc UNION ALL select the_geom, HEIGHT from "+building_table_name+" ")
-
-            sql.execute("DROP TABLE IF EXISTS BUFFERED_SCREENS")
-
-
-        }else{
-            sql.execute("DROP TABLE IF EXISTS BUILDINGS_SCREENS")
-            sql.execute("CREATE TABLE BUILDINGS_SCREENS as select ST_SimplifyPreserveTopology(ST_BUFFER(sc.the_geom,0.1, 'join=bevel'),0.1) the_geom, HEIGHT HEIGHT from SCREENS sc")
-
-        }
+        // Merge untruncated screens and truncated screens
+        sql.execute("DROP TABLE IF EXISTS TMP_SCREENS")
+        sql.execute("create table TMP_SCREENS as select s.pk, s.the_geom, s.height from "+screen_table_name+" s where pk not in (select pk_screen from tmp_screen_truncated) UNION ALL select pk_screen, the_geom, height from tmp_screen_truncated;")
+        sql.execute("drop table if exists tmp_screen_truncated;")
+        // Convert linestring screens to polygons with buffer function
+        sql.execute("DROP TABLE IF EXISTS TMP_BUFFERED_SCREENS")
+        sql.execute("CREATE TABLE TMP_BUFFERED_SCREENS as select ST_BUFFER(sc.the_geom,0.1, 'join=mitre endcap=flat') the_geom,  HEIGHT HEIGHT from TMP_SCREENS sc")
+        sql.execute("DROP TABLE IF EXISTS TMP_SCREENS")
+        // Merge buildings and buffered screens
+        sql.execute("DROP TABLE IF EXISTS BUILDINGS_SCREENS")
+        sql.execute("CREATE TABLE BUILDINGS_SCREENS as select the_geom the_geom, HEIGHT from TMP_BUFFERED_SCREENS sc UNION ALL select the_geom, HEIGHT from "+building_table_name+" ")
+        sql.execute("DROP TABLE IF EXISTS TMP_BUFFERED_SCREENS")
+        sql.execute("DROP TABLE IF EXISTS BUFFERED_SCREENS")
 
 
-        sql.execute("Create spatial index on BUILDINGS_SCREENS(the_geom);")
-        sql.execute("ALTER TABLE BUILDINGS_SCREENS ADD pk INT AUTO_INCREMENT PRIMARY KEY;" )
-
+    }else{
+        sql.execute("DROP TABLE IF EXISTS BUILDINGS_SCREENS")
+        sql.execute("CREATE TABLE BUILDINGS_SCREENS as select ST_BUFFER(sc.the_geom,0.1, 'join=mitre endcap=flat') the_geom, HEIGHT HEIGHT from "+screen_table_name+" sc")
 
     }
+
+
+    sql.execute("Create spatial index on BUILDINGS_SCREENS(the_geom);")
+    sql.execute("ALTER TABLE BUILDINGS_SCREENS ADD pk INT AUTO_INCREMENT PRIMARY KEY;" )
+
 
     return [tableNameCreated: "Process done ! Table BUILDINGS_SCREENS has been created"]
 }
